backend/requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
Pillow

tensorflow-serving-api
matplotlib
numpy
pydantic

google.generativeai